Migrate from int to int64 for db Add function for getting supplier, product types and ingredients

+++ pkg/models/coordinate.go
@@ -1,7 +1,7 @@
 package models
 
 type Coordinate struct {
-	ID        int
+	ID        int64
 	X, Y      float64
 	Humanized string
 	Name      string
